refactor: make Matcher package self-contained
- Isolated Matcher functionality to remove external dependencies
- Improved encapsulation to enhance modularity and ease of reuse
