+++ SubRenamer/Matcher/Helper.cs
@@ -2,7 +2,6 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text.RegularExpressions;
-using SubRenamer.Helper;
 
 namespace SubRenamer.Matcher;
 
@@ -17,7 +16,7 @@ public static class Helper
         }
         catch (Exception e)
         {
-            Logger.Out.WriteLine(e.Message);
+            Logger.Out.WriteLine("[ExtractMatchKeyRegex] Exception: {0}", e.Message);
         }
 
         return "";
@@ -81,5 +80,66 @@ public static class Helper
         List<MatchItem> result = [..items];
         result.Sort((a, b) => new MixedStringComparer().Compare(a.Key, b.Key));
         return result;
+    }
+    
+    /// <summary>
+    /// Compares two strings using a mixed comparison algorithm.
+    /// </summary>
+    /// <remarks>
+    /// The algorithm compares strings by their numeric and non-numeric parts.
+    /// </remarks>
+    public class MixedStringComparer : IComparer<string>
+    {
+        public int Compare(string? x, string? y)
+        {
+            if (x is null || y is null) return 0;
+            return NaturalCompare(x, y);
+        }
+
+        private int NaturalCompare(string str1, string str2)
+        {
+            int length1 = str1.Length;
+            int length2 = str2.Length;
+            int index1 = 0;
+            int index2 = 0;
+
+            while (index1 < length1 && index2 < length2)
+            {
+                if (char.IsDigit(str1[index1]) && char.IsDigit(str2[index2]))
+                {
+                    int num1 = 0;
+                    int num2 = 0;
+
+                    while (index1 < length1 && char.IsDigit(str1[index1]))
+                    {
+                        num1 = num1 * 10 + (str1[index1] - '0');
+                        index1++;
+                    }
+
+                    while (index2 < length2 && char.IsDigit(str2[index2]))
+                    {
+                        num2 = num2 * 10 + (str2[index2] - '0');
+                        index2++;
+                    }
+
+                    if (num1 != num2)
+                    {
+                        return num1.CompareTo(num2);
+                    }
+                }
+                else
+                {
+                    if (str1[index1] != str2[index2])
+                    {
+                        return str1[index1].CompareTo(str2[index2]);
+                    }
+
+                    index1++;
+                    index2++;
+                }
+            }
+
+            return length1.CompareTo(length2);
+        }
     }
 }
