+++ SubRenamer/Matcher/Logger.cs
